add authorize middle and add requester for note

+++ module/note/notebusiness/delete_note.go
@@ -13,11 +13,12 @@ type DeleteNoteStore interface {
 }
 
 type deleteNoteBiz struct {
-	store DeleteNoteStore
+	store     DeleteNoteStore
+	requester common.Requester
 }
 
-func NewDeleteNoteBiz(store DeleteNoteStore) *deleteNoteBiz {
-	return &deleteNoteBiz{store: store}
+func NewDeleteNoteBiz(store DeleteNoteStore, requester common.Requester) *deleteNoteBiz {
+	return &deleteNoteBiz{store: store, requester: requester}
 }
 
 func (biz *deleteNoteBiz) DeleteNote(ctx context.Context, noteId int) error {
@@ -37,8 +38,11 @@ func (biz *deleteNoteBiz) DeleteNote(ctx context.Context, noteId int) error {
 	}
 
 	if note.Status == 0 {
-		//return common.ErrCannotDeleteEntity()
-		return common.ErrDB(errors.New("note has been deleted before"))
+		return common.ErrCannotDeleteEntity(notemodel.NoteTableName, errors.New("note has been deleted before"))
+	}
+
+	if note.UserId != biz.requester.GetUserId() {
+		return common.ErrNoPermission(errors.New("you are not owner"))
 	}
 
 	if err := biz.store.Delete(ctx, note.Id); err != nil {
